backend/foodgram/requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-filter==22.1
django-import-export==2.8.0
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-filters==1.0.0.dev2
djangorestframework-simplejwt==4.7.0
djoser==2.1.0
drf-extra-fields==3.4.1
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.21.6
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tablib==3.2.1
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
